Updated groupdatevolume function.
Updated groupdatevolume function to add the ability to override the wordcloud2 minSize parameter when requesting a wordcloud (see README).

Also added retry logic to the functions groupdatevolume calls to collect sample set dates and  optional subject fields.  I encountered exceptions when the network was experiencing some load where the python call wasn't returning a proper response to the "listheaderfield" call.  When the function tried to access the desired part of the response, it would end with a "subscript out of range".  I found that retrying the call eventually returned a proper response.  I added code to retry the call (up to 10 times) until there's a reponse with with correct structure.