Added proper exception handling for nntplib call, removed timeout option on header list calls.
After encountering a few nntp exceptions in testing the package, I decided that proper exception handling was overdue.  It isn't perfect, but it's 100% better than the previous nntpr release.

I had a timeout option working with listartheader and listheaderfield, where it would interupt the python nntplib processing after a specified number of seconds.  This had to be removed, the timeout exceptions caused at the R level and the python level exception catching when running nntp calls through reticulate didn't function properly.  I didn't have a good replacement for the R timeout in python 2.7, but it's something I plan to revisit down the road.

Version 1.9.0